requirements-test.txt
# Testing Dependencies for DogDating Project
# ===========================================
# Install with: pip install -r requirements-test.txt
# Or for development: pip install -r requirements.txt -r requirements-test.txt

# Core Testing Framework
pytest>=7.4.0
pytest-django>=4.5.2
pytest-cov>=4.1.0
pytest-xdist>=3.3.1          # Parallel test execution
pytest-sugar>=0.9.7          # Better test output
pytest-timeout>=2.1.0        # Test timeouts
pytest-mock>=3.11.1          # Mocking utilities

# Test Data Generation
factory-boy>=3.3.0
Faker>=19.0.0

# Coverage Reporting
coverage[toml]>=7.3.0

# Code Quality
flake8>=6.1.0
black>=23.9.0
isort>=5.12.0
mypy>=1.5.0

# Django Testing Utilities
django-debug-toolbar>=4.2.0

# Optional: Performance Testing
pytest-benchmark>=4.0.0

# Optional: Database Testing
# pytest-postgresql>=5.0.0  # Removed due to psycopg dependency issues